+++ web/src/collection/components/ProcessingPage/TaskSidebar/index.js
@@ -1,0 +1,1 @@
+export { default } from "./TaskSidebar";
